fix circular withMembership dependency
